requirements.txt
Flask==2.1.1
PyJWT==2.4.0
pyparsing==3.0.6
PyQt5==5.15.6
requests==2.26.0
storm==0.25
werkzeug==2.0.3
